testing: migrate to trampoline_v2 (#3860)
* testing: prototype trampoline_v2

* add RUN_TESTS_DIRS envvar

* add a script for running tests locally

* don't rely on git, remove noxfile.py only if we copied it

* add TRAMPOLINE_IMAGE_UPLOAD envvar

* add example

* add .trampolinerc_tmpl

* allow docker in docker

* use /workspace for project root

* use $PROJECT_ROOT instead of $PWD

* also pass down KOKORO_KEYSTORE_DIR

* pass down Kokoro dynamic variables

* introduce TRAMPOLINE_CI and TRAMPOLINE_V2 envvar

* use --entrypoint instead of giving the command

* add kbuilder's uid as the default arg

* do not show the docker commands by default

* make the mount point configurable

* change TRAMPOLINE_IMAGE_SOURCE to TRAMPOLINE_DOCKERFILE

* only decrypt the secret when the key exists

* also prevents decrypt-secrets.sh to override existing files.

* remove secret if we used decrypt-secrets.sh before the test

* add 2 more envvars for build cop bot

* setup kokoro service account only on kokoro

* address review comments

* correctly report the exit value

* remove unnecessary build args

* use docker gid

* download cloud_sql_proxy to ${HOME} again

+++ .kokoro/tests/run_tests.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-
 # Copyright 2019 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -32,23 +31,29 @@ if [[ $* == *--only-diff-head* ]]; then
     DIFF_FROM="HEAD~.."
 fi
 
-cd github/python-docs-samples
+if [[ -z "${PROJECT_ROOT:-}" ]]; then
+    PROJECT_ROOT="github/python-docs-samples"
+fi
+
+cd "${PROJECT_ROOT}"
+
+# add user's pip binary path to PATH
+export PATH="${HOME}/.local/bin:${PATH}"
 
 # install nox for testing
-pip install -q nox
+pip install --user -q nox
 
-# Use secrets acessor service account to get secrets
+# Use secrets acessor service account to get secrets.
 if [[ -f "${KOKORO_GFILE_DIR}/secrets_viewer_service_account.json" ]]; then
     gcloud auth activate-service-account \
 	   --key-file="${KOKORO_GFILE_DIR}/secrets_viewer_service_account.json" \
 	   --project="cloud-devrel-kokoro-resources"
+    # This script will create 3 files:
+    # - testing/test-env.sh
+    # - testing/service-account.json
+    # - testing/client-secrets.json
+    ./scripts/decrypt-secrets.sh
 fi
-
-# This script will create 3 files:
-# - testing/test-env.sh
-# - testing/service-account.json
-# - testing/client-secrets.json
-./scripts/decrypt-secrets.sh
 
 source ./testing/test-env.sh
 export GOOGLE_APPLICATION_CREDENTIALS=$(pwd)/testing/service-account.json
@@ -63,9 +68,12 @@ export GOOGLE_CLIENT_SECRETS=$(pwd)/testing/client-secrets.json
 export DATALABELING_ENDPOINT="test-datalabeling.sandbox.googleapis.com:443"
 
 # Run Cloud SQL proxy (background process exit when script does)
-wget --quiet https://dl.google.com/cloudsql/cloud_sql_proxy.linux.amd64 -O cloud_sql_proxy && chmod +x cloud_sql_proxy
-./cloud_sql_proxy -instances="${MYSQL_INSTANCE}"=tcp:3306 &>> cloud_sql_proxy.log &
-./cloud_sql_proxy -instances="${POSTGRES_INSTANCE}"=tcp:5432 &>> cloud_sql_proxy.log &
+wget --quiet https://dl.google.com/cloudsql/cloud_sql_proxy.linux.amd64 \
+     -O ${HOME}/cloud_sql_proxy && chmod +x ${HOME}/cloud_sql_proxy
+${HOME}/cloud_sql_proxy -instances="${MYSQL_INSTANCE}"=tcp:3306 &>> \
+       ${HOME}/cloud_sql_proxy.log &
+${HOME}/cloud_sql_proxy -instances="${POSTGRES_INSTANCE}"=tcp:5432 &>> \
+       ${HOME}/cloud_sql_proxy-postgres.log &
 echo -e "\nCloud SQL proxy started."
 
 echo -e "\n******************** TESTING PROJECTS ********************"
@@ -92,6 +100,36 @@ for file in **/requirements.txt; do
     file=$(dirname "$file")
     cd "$file"
 
+    # First we look up the environment variable `RUN_TESTS_DIRS`. If
+    # the value is set, we'll iterate through the colon separated
+    # directory list. If the target directory is not under any
+    # directory in the list, we skip this directory.
+    # This environment variables are primarily for
+    # `scripts/run_tests_local.sh`.
+    #
+    # The value must be a colon separated list of relative paths from
+    # the project root.
+    #
+    # Example:
+    #   cdn:appengine/flexible
+    #     run tests for `cdn` and `appengine/flexible` directories.
+    #   logging/cloud-client
+    #     only run tests for `logging/cloud-client` directory.
+    #
+    if [[ -n "${RUN_TESTS_DIRS:-}" ]]; then
+	match=0
+	for d in $(echo "${RUN_TESTS_DIRS}" | tr ":" "\n"); do
+	    # If the current dir starts with one of the
+	    # RUN_TESTS_DIRS, we should run the tests.
+	    if [[ "${file}" = "${d}"* ]]; then
+		match=1
+		break
+	    fi
+	done
+	if [[ $match -eq 0 ]]; then
+	    continue
+	fi
+    fi
     # If $DIFF_FROM is set, use it to check for changes in this directory.
     if [[ -n "${DIFF_FROM:-}" ]] && [[ "${test_all}" == "false" ]]; then
         git diff --quiet "$DIFF_FROM" .
@@ -108,24 +146,27 @@ for file in **/requirements.txt; do
 
     # If no local noxfile exists, copy the one from root
     if [[ ! -f "noxfile.py" ]]; then
-      PARENT_DIR=$(cd ../ && pwd)
-      while [[ "$PARENT_DIR" != "$ROOT" && ! -f "$PARENT_DIR/noxfile-template.py" ]];
-      do
-        PARENT_DIR=$(dirname "$PARENT_DIR")
-      done
-      cp "$PARENT_DIR/noxfile-template.py" "./noxfile.py"
-      echo -e "\n Using noxfile-template from parent folder ($PARENT_DIR). \n"
+        PARENT_DIR=$(cd ../ && pwd)
+        while [[ "$PARENT_DIR" != "$ROOT" && ! -f "$PARENT_DIR/noxfile-template.py" ]];
+        do
+            PARENT_DIR=$(dirname "$PARENT_DIR")
+        done
+        cp "$PARENT_DIR/noxfile-template.py" "./noxfile.py"
+        echo -e "\n Using noxfile-template from parent folder ($PARENT_DIR). \n"
+        cleanup_noxfile=1
+    else
+        cleanup_noxfile=0
     fi
 
     # Use nox to execute the tests for the project.
     nox -s "$RUN_TESTS_SESSION"
     EXIT=$?
 
-    # If REPORT_TO_BUILD_COP_BOT is set to "True", send the test log
+    # If REPORT_TO_BUILD_COP_BOT is set to "true", send the test log
     # to the Build Cop Bot.
     # See:
     # https://github.com/googleapis/repo-automation-bots/tree/master/packages/buildcop.
-    if [[ "${REPORT_TO_BUILD_COP_BOT:-}" == "True" ]]; then
+    if [[ "${REPORT_TO_BUILD_COP_BOT:-}" == "true" ]]; then
       chmod +x $KOKORO_GFILE_DIR/linux_amd64/buildcop
       $KOKORO_GFILE_DIR/linux_amd64/buildcop
     fi
@@ -137,10 +178,17 @@ for file in **/requirements.txt; do
       echo -e "\n Testing completed.\n"
     fi
 
+    # Remove noxfile.py if we copied.
+    if [[ $cleanup_noxfile -eq 1 ]]; then
+        rm noxfile.py
+    fi
+
 done
 cd "$ROOT"
 
-# Workaround for Kokoro permissions issue: delete secrets
-rm testing/{test-env.sh,client-secrets.json,service-account.json}
+# Remove secrets if we used decrypt-secrets.sh.
+if [[ -f "${KOKORO_GFILE_DIR}/secrets_viewer_service_account.json" ]]; then
+    rm testing/{test-env.sh,client-secrets.json,service-account.json}
+fi
 
 exit "$RTN"
